Add separate headers for platform specific code
