+++ src/runtime_src/core/common/message.h
@@ -1,18 +1,6 @@
-/**
- * Copyright (C) 2016-2021 Xilinx, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may
- * not use this file except in compliance with the License. A copy of the
- * License is located at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
+// SPDX-License-Identifier: Apache-2.0
+// Copyright (C) 2016-2022 Xilinx, Inc. All rights reserved.
+// Copyright (C) 2022-2026 Advanced Micro Devices, Inc. All rights reserved.
 
 #ifndef xrtcore_message_h_
 #define xrtcore_message_h_
@@ -27,6 +15,14 @@
 namespace xrt_core { namespace message {
 
 using severity_level = xrt::message::level;
+
+class message_dispatch
+{
+public:
+  virtual ~message_dispatch() = default;
+  static message_dispatch* make_dispatcher(const std::string& choice);
+  virtual void send(severity_level l, const char* tag, const char* msg) = 0;
+};
 
 XRT_CORE_COMMON_EXPORT
 void
